+++ ts-client/src/amm/types/index.ts
@@ -11,8 +11,6 @@ export type AmmProgram = Program<AmmIdl>;
 export type VaultProgram = Program<VaultIdl>;
 
 export interface AmmImplementation {
-  tokenA: TokenInfo;
-  tokenB: TokenInfo;
   decimals: number;
   isStablePool: boolean;
   updateState: () => Promise<void>;
@@ -37,6 +35,11 @@ type Fees = {
   lp?: BN;
   tokenA: BN;
   tokenB: BN;
+};
+
+export type tokenAddressAndDecimals = {
+  address: string;
+  decimals: number;
 };
 
 export interface LockEscrow {
@@ -113,7 +116,7 @@ export type StableSwapCurve = {
     amp: BN;
     tokenMultiplier: TokenMultiplier;
     depeg: Depeg;
-    lastAmpUpdatedTimestamp: BN,
+    lastAmpUpdatedTimestamp: BN;
   };
 };